Import explicit modules
Importing explicit modules reduced the amount of object held in memory as consequence of importing from index.ts files
