+++ libs/ui-lib/lib/common/api/utils.ts
@@ -1,7 +1,7 @@
 import Axios, { AxiosError } from 'axios';
 import * as Sentry from '@sentry/browser';
 import pick from 'lodash-es/pick.js';
-import {
+import type {
   Error as APIError,
   InfraError,
 } from '@openshift-assisted/types/assisted-installer-service';
