+++ libs/ui-lib/lib/common/index.ts
@@ -9,3 +9,4 @@ export * from './config';
 export * from './features';
 export * from './reducers';
 export * from './selectors';
+export { ResourceUIState } from './types/resource-ui-state';
