+++ libs/ui-lib/lib/common/types/resource-ui-state.ts
@@ -1,0 +1,8 @@
+export enum ResourceUIState {
+  LOADING = 'LOADING',
+  RELOADING = 'RELOADING',
+  EMPTY = 'EMPTY',
+  LOADED = 'LOADED',
+  POLLING_ERROR = 'POLLING_ERROR',
+  UPDATE_ERROR = 'UPDATE_ERROR',
+}
