+++ libs/ui-lib/lib/ocm/store/slices/feature-flags/selectors.ts
@@ -5,10 +5,8 @@ import { createSelector } from '@reduxjs/toolkit';
 export const selectFeatureFlagsSlice = (state: RootStateDay1) => state.featureFlags;
 
 export const isFeatureEnabled = createSelector(
-  [
-    selectFeatureFlagsSlice,
-    (_featureFlags: ReturnType<typeof selectFeatureFlagsSlice>, featureId: keyof FeatureListType) =>
-      featureId,
-  ],
+  selectFeatureFlagsSlice,
+  (_featureFlags: ReturnType<typeof selectFeatureFlagsSlice>, featureId: keyof FeatureListType) =>
+    featureId,
   (featureFlags, featureId) => featureFlags.data[featureId],
 );
